Refactor code structure for Virtual Ticket
