+++ src/pages/Results/components/VirtualTicket/shared/hooks.ts
