+++ src/pages/Results/pages/FootO/pages/StartTime/FootOStartTime.tsx
@@ -6,7 +6,7 @@ import {useTranslation} from "react-i18next";
 import ResultListContainer from "../../components/ResultListContainer.tsx";
 import ResultListItem from "../../components/ResultListItem.tsx";
 import StartTime from '../../../../components/StartTime.tsx'
-import {ProcessedRunnerModel} from "../../../shared/EntityTypes.ts";
+import {ProcessedRunnerModel} from "../../../../components/VirtualTicket/shared/EntityTypes.ts";
 
 export default function FootOStartTime() {
   const {t} = useTranslation()
